requirements.txt
mcp
fastmcp
requests
python-dotenv
pillow
ffmpeg-python
